exercicio_06/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
scikit-learn==1.4.2
numpy==1.26.4
joblib==1.4.2
Pillow==10.3.0
python-multipart==0.0.9
tensorflow==2.16.1
streamlit==1.35.0
requests==2.32.3
httpx==0.27.0
pytest==8.2.0
